% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_frame.R
\name{Class_frame}
\alias{Class_frame}
\title{Class_frame}
\usage{
Class_frame(dataframe.seq, n)
}
\arguments{
\item{dataframe.seq}{is the dataframe converted by Convert_seq funciton}

\item{n}{is the lenght of sequence, if the target sequecne is less than
n, it will belong to Hotspot, otherwise is Exome}
}
\value{
list
}
\description{
Based on dataframe converted by Convert_seq funciton, this function can
divide the target seqence into Hotspot and Exome two list by thier sequence length
}
\examples{
Chr<-c("chrX","chrX" ,"chrX","chrX" )
Target.Start<-c(153996577,153999034,154004462,154002877)
Target.End<-c(153996707 ,153999154,154004599,154002980)
Strand<-rep("+",4)
chromosome <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
frame<-Convert_seq(Chr,Target.Start,Target.End,Strand,chromosome)
class<-Class_frame(frame,125)
sapply(class,nrow)
}
