% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_seq.R
\name{Convert_seq}
\alias{Convert_seq}
\title{Convert_seq}
\usage{
Convert_seq(Chr, Target.Start, Target.End, Strand, chromosome, NO. = NA)
}
\arguments{
\item{Chr}{Chromosome information}

\item{Target.Start}{Starting location of target sequence}

\item{Target.End}{ending ocation of target sequence}

\item{Strand}{+/- strand}

\item{chromosome}{chromosome information
e.g chromosome <- BSgenome.Hsapiens.UCSC.hg19}

\item{NO.}{name of the each target sequence;default is NULL}
}
\value{
a dataframe
}
\description{
You can extract the targe sequence and convert it into dataframe
}
\examples{
Chr<-c("chrX","chrX" ,"chrX","chrX" )
Target.Start<-c(153996577,153999034,154004462,154002877)
Target.End<-c(153996707 ,153999154,154004599,154002980)
Strand<-rep("+",4)
chromosome <- BSgenome.Hsapiens.UCSC.hg19
Convert_seq(Chr,Target.Start,Target.End,Strand,chromosome)

}
