% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exo_extension.R
\name{Exo_extension}
\alias{Exo_extension}
\title{Exo_extension}
\usage{
Exo_extension(EXOME, n, cut.length, chromosome)
}
\arguments{
\item{EXOME}{is dataframe converted by Convert_seq funciton and need to be sheared}

\item{n}{total lenght of target sequence after amplifying}

\item{cut.length}{Clip length of target sequecne, the sheared fragment is reuqired to be less than "cut.length"
#e.g. for 345bp length of target sequence, it can be sheared into 1-178 and 179-354 two fragments}

\item{chromosome}{is the chromosome information for amplified fragment
e.g chromosome <- BSgenome.Hsapiens.UCSC.hg19}
}
\description{
For the target sequecne dataframe, if you  need to shear, you can directly use Exo_extension function,
which not only shears the target sequence,but also amplifies the fragments for designing the primer. Finally,
sequence of sheared fragment and amplified fragment will be returned to the original dataframe, respectively
called Fragment.seq and Extend.fragment. Meanwhile, it additionally show the location and length information of sheared
fragments. Of course，each of complete amplifyed fragment sequence will also showed at dataframe, called Extend.seq.
}
\examples{
Chr<-c("chrX","chrX" ,"chrX","chrX" )
Target.Start<-c(153996577,153999034,154004462,154002877)
Target.End<-c(153996707 ,153999154,154004599,154002980)
Strand<-rep("+",4)
chromosome <- BSgenome.Hsapiens.UCSC.hg19：：BSgenome.Hsapiens.UCSC.hg19
frame<-Convert_seq(Chr,Target.Start,Target .End,Strand,chromosome)
class<-Class_frame(frame,125)
sapply(class,nrow)
chromosome <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
entend.exome<-Exo_extension(class$EXOME,280,200,chromosome)
View(entend.exome)
}
