% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Inf.R
\name{Filter_Inf}
\alias{Filter_Inf}
\title{Filter_Inf}
\usage{
Filter_Inf(hot.pair, plist, distance, overlap, inf = TRUE)
}
\arguments{
\item{hot.pair}{is the primer pair}

\item{plist}{is the list formation of  Primers under each target sequence}

\item{distance}{threshold for distance length between primer pairs, the distacne between primer pairs reqiure
to be less than the threshold}

\item{overlap}{threshold for ovelaps length between primer pairs,the overlap between primer pairs reqiure
to be less than the threshold}
}
\description{
Filter_Inf not only filter the primer pairs, but also extract the basical information for each primer
}
\examples{
plist<-List_PrimerSet(hot.out)
hot.pair<-List_PairSet(plist)
Filter_Inf(hot.pair,plist,distance=20,overlap=15)

#Don't return information of primer
Filter_Inf(hot.pair,plist,distance=20,overlap=15,inf=FALSE)
}
