% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List_PairExomeSet.R
\name{List_PairExomeSet}
\alias{List_PairExomeSet}
\title{List_PairExomeSet}
\usage{
List_PairExomeSet(eplist, dir = "left")
}
\arguments{
\item{eplist}{is the list containing Exome sheared fragments and  corresponding primers achieved by 'List_PrimerExomeSet' function}

\item{dir}{is the direction of extension of target sequence, if the strand of target sequence
is "+", the direction of extension of target sequence will be "left"; The direction of
"-" strand is "right". sort order of GSP1 and GSP2 depends on dir' parameter. Here, default is "left".}
}
\value{
list
}
\description{
This functions is specifically used for matching primer pair under each sheared fragement of Exome sequene and
the high quality primers are ranked at top
}
\examples{
library(xxx)
data(exo.out)
data(exo)
sp.ex<-Split_exome(exo.out,frame=exo)
eplist<-List_PrimerExomeSet(sp.ex)
List_PairExomeSet(eplist)
List_PairExomeSet(eplist,dir="right")
}
