% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List_PairSet.R
\name{List_PairSet}
\alias{List_PairSet}
\title{List_PairSet}
\usage{
List_PairSet(plist, dir = "left")
}
\arguments{
\item{plist}{is the list formation of  Primers under each target sequence.Thereinto,
GSP1 is the gene specifical primer1, which is target primer and far from extension
sequence；the other closer primer is GSP2.}

\item{dir}{is the direction of extension of target sequence, if the strand of target sequence
is "+", the direction of extension of target sequence will be "left"; The direction of
"-" strand is "right". sort order of GSP1 and GSP2 depends on dir' parameter. Here, default is "left".}
}
\value{
list
}
\description{
List_PairSet can match primer pair, the high quality primers are ranked at top
}
\examples{
plist<-List_PrimerSet(hot.out)
List_PairSet(plist)
List_PairSet(plist,dir="right")
}
