% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List_PrimerExomeSet.R
\name{List_PrimerExomeSet}
\alias{List_PrimerExomeSet}
\title{List_PrimerExomeSet}
\usage{
List_PrimerExomeSet(split.exome)
}
\arguments{
\item{split.exome}{is the list converted from Exome Primer3 result by 'Spliy_exome' function
Thereinto, GSP1 is the gene specifical primer1, which is target primer and far from extension
sequence；the other closer primer is GSP2}
}
\value{
list
}
\description{
This function is used to integrate sheared fragements and corresponding primers into list formation
}
\examples{
sp.ex<-Split_exome(exo.out,frame=exo)
x<-List_PrimerExomeSet(sp.ex)
sapply(x,function(x){c(names(x[1]),names(x[[1]]))})
only for one target
y<-List_PrimerSet(sp.ex[[1]])
sapply(y,names)
}
