% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List_PrimerSet.R
\name{List_PrimerSet}
\alias{List_PrimerSet}
\title{List_PrimerSet}
\usage{
List_PrimerSet(x)
}
\arguments{
\item{x}{is the output result of Primer3}
}
\value{
list
}
\description{
List_PrimerSet
}
\examples{
list<-List_PrimerSet(hot.out)
sapply(list,names)
}
