% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_primer3.R
\name{param_primer3}
\alias{param_primer3}
\title{param_primer3}
\usage{
param_primer3(x, add)
}
\arguments{
\item{x}{is the Priemr3 input data}

\item{add}{supplementary parameters for Primer3}
}
\description{
Due to there are many parameters required by Primer3, you can specifically supplement
the parameters you expected
}
\examples{
chromosome <- BSgenome.Hsapiens.UCSC.hg19
entend.hot<-Hot_extension(hot,185,chromosome)
p.hot<-Primer3_Hotset(entend.hot,return=2)
View(p.hot)
add<-c("PRIMER_INTERNAL_MAX_TM=63","PRIMER_MAX_END_GC=5")
param_primer3(p.hot,add)
}
