% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split_exome.R
\name{Split_exome}
\alias{Split_exome}
\title{Split_exome}
\usage{
Split_exome(exo.out, frame)
}
\arguments{
\item{exo.out}{is the output exome result of Primer3}

\item{frame}{is the exome targe sequence dataframe}
}
\value{
list
}
\description{
Split_exome
}
\examples{
library(xxx)
data(exo.out)
data(exo)
sp.ex<-Split_exome(exo.out,frame=exo)
}
