% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_primer.R
\name{Target_primer}
\alias{Target_primer}
\title{Target_primer}
\usage{
Target_primer(x)
}
\arguments{
\item{x}{List of Primers under each target sequence converted by 'List_PrimerSet' function}
}
\value{
data.frame
}
\description{
This fucntion is used to get information of each primer under the target sequecne
}
\examples{
plist<-List_PrimerSet(hot.out)
Target_primer(plist[[1]])
}
