% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetSet_primer.R
\name{TargetSet_primer}
\alias{TargetSet_primer}
\title{TargetSet_primer}
\usage{
TargetSet_primer(x)
}
\arguments{
\item{x}{List set of Primers under each target sequence converted by 'List_PrimerSet' function}
}
\value{
dataframe
}
\description{
TargetSet_primer
}
\examples{
plist<-List_PrimerSet(hot.out)
TargetSet_primer(plist)

exome
sp.ex<-Split_exome(exo.out,frame=exo)
eplist<-List_PrimerExomeSet(sp.ex)
lapply(eplist,TargetSet_primer) \%>\% list.rbind()
}
