% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualGenome.R
\name{visualGenome}
\alias{visualGenome}
\title{visualGenome}
\usage{
visualGenome(Chr, Target.Start, Target.End, Extend.length, ID, x,
  plot = TRUE)
}
\arguments{
\item{Chr}{Chromosome information，e.g "Chr1"}

\item{Target.Start}{Starting location of target sequence}

\item{Target.End}{ending location of target sequence}

\item{Extend.length}{toal length of extended target sequence}

\item{ID}{Name or NO. of target sequence}

\item{x}{a pair primer 3 avliable from Hot_output or Exome_outpur function}

\item{plot}{plot the genomic location, defaulr is to ‘TRUE’}
}
\value{
GRange of target and primer pair information
}
\description{
visualGenome
}
\examples{
Chr<-"chr5"
Target.Start<-176939369
Target.End<-176939371
Extend.length<-185
ID<-"leukemia_36"
plist<- List_PrimerSet(hot.out)
hot.pair<-List_PairSet(plist)
rem_Pairs.inf<-Filter_Inf(hot.pair,plist,distance=20,overlap=10)
x<-Hot_output(rem_Pairs.inf,Write.out=FALSE)[3,]
visualGenome(Chr,Target.Start,Target.End,Extend.length,ID,x)
#only return local information
visualGenome(Chr,Target.Start,Target.End,Extend.length,ID,x,plot=FALSE)
}
